separate types tests
